requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
sentence-transformers
faiss-cpu
openai
slowapi
python-dotenv
python-multipart
httpx
